Dungeon moved in separate ctalog
- Rectangle private Map removed
